+++ src/app/components/ApplyPage/FAQAccordion/FAQAccordion.scss
@@ -10,7 +10,6 @@
     max-width: 900px;
     margin: 4rem auto;
     color: #000;
-    font-family: 'Inter', sans-serif;
   }
 
   .AccordionItem {
@@ -67,6 +66,7 @@
     padding-left: 3rem;
     padding-top: 1rem;
     font-size: 1.5rem;
+    font-family: 'Inter', sans-serif;
     color: #333;
     line-height: 1.6;
     max-height: 0;
