+++ src/app/components/ApplyPage/PositionComponent/PositionComponent.tsx
@@ -1,0 +1,3 @@
+export default function Position() {
+  return <main></main>;
+}
