+++ src/app/components/ApplyPage/PositionComponent/utils.ts
